.ipynb_checkpoints/requirements-checkpoint.txt
botbuilder-integration-aiohttp>=4.15.0
azure-ai-textanalytics

